requirements.txt
Adafruit-Blinka==8.10.0
adafruit-circuitpython-busdevice==5.2.3
adafruit-circuitpython-neopixel==6.3.7
adafruit-circuitpython-pixelbuf==2.0.0
adafruit-circuitpython-requests==1.12.11
adafruit-circuitpython-typing==1.8.3
Adafruit-PlatformDetect==3.38.0
Adafruit-PureIO==1.1.9
adafruit-python-shell==1.3.3
appdirs==1.4.4
arandr==0.1.10
args==0.1.0
astroid==2.5.1
asttokens==2.0.4
automationhat==0.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
blinker==1.4
blinkt==0.1.2
buttonshim==0.0.2
Cap1xxx==0.1.3
certifi==2020.6.20
chardet==4.0.0
click==7.1.2
clint==0.5.1
colorama==0.4.4
colorzero==1.1
cryptography==3.3.2
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==4.4.2
distro==1.5.0
docutils==0.16
drumhat==0.1.0
envirophat==1.0.0
ExplorerHAT==0.4.2
Flask==1.1.2
fourletterphat==0.1.0
gpiozero==1.6.2
guizero==1.1.1
html5lib==1.1
idna==2.10
ipykernel==5.4.3
ipython==7.20.0
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
jupyter-client==6.1.11
jupyter-core==4.7.1
kiwisolver==1.3.1
lazy-object-proxy==0.0.0
logilab-common==1.8.1
lxml==4.6.3
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
microdotphat==0.2.1
mock==4.0.3
mote==0.0.4
motephat==0.0.3
mypy==0.812
mypy-extensions==0.4.3
nudatus==0.0.5
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
pantilthat==0.0.7
parso==0.8.1
pbr==5.5.0
pexpect==4.8.0
pgzero==1.2
phatbeat==0.1.1
pianohat==0.1.0
picamera==1.13
pickleshare==0.7.5
piglow==1.2.5
pigpio==1.78
Pillow==8.1.2
prompt-toolkit==3.0.14
psutil==5.8.0
pycairo==1.16.2
pycodestyle==2.6.0
pycups==2.0.1
pyflakes==2.2.0
pyftdi==0.54.0
pygame==1.9.6
Pygments==2.7.1
PyGObject==3.38.0
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.7.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pyserial==3.5
pysmbc==1.0.23
python-apt==2.2.1
python-dateutil==2.8.1
pyusb==1.2.1
pyzmq==20.0.0
QScintilla==2.11.6
qtconsole==5.0.2
QtPy==1.9.0
rainbowhat==0.1.0
reportlab==3.5.59
requests==2.25.1
requests-oauthlib==1.0.0
responses==0.12.1
roman==2.0.0
rpi-ws281x==4.3.4
RPi.GPIO==0.7.1
RTIMULib==7.2.1
SCons==4.0.1
scrollphat==0.0.7
scrollphathd==1.2.1
semver==2.10.2
Send2Trash==1.6.0b1
sense-hat==2.4.0
simplejson==3.17.2
six==1.16.0
skywriter==0.0.7
sn3218==1.2.7
soupsieve==2.2.1
spidev==3.5
ssh-import-id==5.10
sysv-ipc==1.1.0
thonny==3.3.14
toml==0.10.1
tornado==6.1
touchphat==0.0.1
traitlets==5.0.5
twython==3.8.2
typed-ast==1.4.2
typing-extensions==4.4.0
uflash==1.2.4
unicornhathd==0.0.4
urllib3==1.26.5
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1